Get rid of strategies module
