+++ src/simulation/investment_tools.rs
@@ -1,4 +1,3 @@
 pub mod constraints;
 pub mod costs;
 pub mod optimisation;
-pub mod strategies;
